+++ integrations/watermilllog/logger.go
@@ -24,22 +24,49 @@ package watermilllog
 import (
 	"github.com/ThreeDotsLabs/watermill"
 	"github.com/goph/logur"
+	"github.com/pkg/errors"
 )
+
+// ErrorHandler handles an error passed to the logger.
+type ErrorHandler interface {
+	// Handle handles an error passed to the logger.
+	Handle(err error, fields map[string]interface{})
+}
 
 // Logger is a github.com/ThreeDotsLabs/watermill.LoggerAdapter compatible logger.
 type Logger struct {
-	logger logur.Logger
+	logger       logur.Logger
+	errorHandler ErrorHandler
 }
 
 // New returns a github.com/ThreeDotsLabs/watermill.LoggerAdapter compatible logger.
-func New(l logur.Logger) *Logger {
-	return &Logger{l}
+func New(logger logur.Logger) *Logger {
+	return &Logger{
+		logger: logger,
+	}
+}
+
+// NewWithErrorHandler returns a github.com/ThreeDotsLabs/watermill.LoggerAdapter compatible logger.
+// Compared to the logger returned by New, this logger sends errors to the error handler.
+func NewWithErrorHandler(logger logur.Logger, errorHandler ErrorHandler) *Logger {
+	return &Logger{
+		logger:       logger,
+		errorHandler: errorHandler,
+	}
 }
 
 func (l *Logger) Error(msg string, err error, fields watermill.LogFields) {
-	fields["err"] = err
+	if l.errorHandler == nil {
+		fields["err"] = err
 
-	l.logger.Error(msg, fields)
+		l.logger.Error(msg, fields)
+
+		return
+	}
+
+	err = errors.WithMessage(err, msg)
+
+	l.errorHandler.Handle(err, fields)
 }
 
 func (l *Logger) Info(msg string, fields watermill.LogFields) {
@@ -59,5 +86,5 @@ func (l *Logger) With(fields watermill.LogFields) watermill.LoggerAdapter {
 		return l
 	}
 
-	return New(logur.WithFields(l.logger, fields))
+	return NewWithErrorHandler(logur.WithFields(l.logger, fields), l.errorHandler)
 }